+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # Changelog
 
 ## 0.9.0 (2019-06-30)
-* NoteTab public release
+* Jupytab public release